% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coalesce.R
\name{coalescent}
\alias{coalescent}
\title{Coalescent simulation}
\usage{
coalescent(n = 10, theta = 1, newick = FALSE)
}
\arguments{
\item{n}{number of samples}

\item{theta}{per generation mutation rate}

\item{newick}{logical; return newick string representation of tree}
}
\value{
vector containing TMRCA, TBL, S_n, and possibly newick
    representation of tree
}
\description{
Coalescent simulation
}

