% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/watterson.R
\name{estimate_watterson}
\alias{estimate_watterson}
\title{watterson estimator for several fragments}
\usage{
estimate_watterson(x = 10, n = 100, L = 200, theta = 0.01)
}
\arguments{
\item{int}{x number of fragments}

\item{int}{n sample size}

\item{int}{L fragment size}

\item{double}{theta per site mutation rate}
}
\value{
theta_w for each fragment
}
\description{
watterson estimator for several fragments
}

