% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/watterson.R
\name{watterson}
\alias{watterson}
\title{watterson estimator}
\usage{
watterson(Sn, n)
}
\arguments{
\item{int}{Sn number of segregating sites}

\item{int}{n number of samples}
}
\value{
theta_w
}
\description{
watterson estimator
}

