% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coalesce.R
\name{coalescent}
\alias{coalescent}
\title{Coalescent simulation}
\usage{
coalescent(n = 10, theta = 1)
}
\arguments{
\item{int}{n number of samples}

\item{double}{theta mutation rate}
}
\value{
tree coalescent tree
}
\description{
Coalescent simulation
}

